Moved DTO entities to their own file
